Rename AbpExtraPropertyDictionaryJsonConverter to AbpHasExtraPropertiesJsonConverter.
Only for private or protected IHasExtraProperties' ExtraProperties.
