+++ framework/src/Volo.Abp.AspNetCore.Mvc/Volo/Abp/AspNetCore/Mvc/Json/AbpJsonOptionsSetup.cs
@@ -29,7 +29,7 @@ namespace Volo.Abp.AspNetCore.Mvc.Json
             options.JsonSerializerOptions.Converters.Add(new AbpStringToBooleanConverter());
 
             options.JsonSerializerOptions.Converters.Add(new ObjectToInferredTypesConverter());
-            options.JsonSerializerOptions.Converters.Add(new AbpExtraPropertyDictionaryJsonConverterFactory());
+            options.JsonSerializerOptions.Converters.Add(new AbpHasExtraPropertiesJsonConverterFactory());
         }
     }
 }
