+++ framework/src/Volo.Abp.Json/Volo/Abp/Json/SystemTextJson/AbpSystemTextJsonSerializerOptionsSetup.cs
@@ -23,7 +23,7 @@ namespace Volo.Abp.Json.SystemTextJson
             options.JsonSerializerOptions.Converters.Add(new AbpStringToBooleanConverter());
 
             options.JsonSerializerOptions.Converters.Add(new ObjectToInferredTypesConverter());
-            options.JsonSerializerOptions.Converters.Add(new AbpExtraPropertyDictionaryJsonConverterFactory());
+            options.JsonSerializerOptions.Converters.Add(new AbpHasExtraPropertiesJsonConverterFactory());
         }
     }
 }
